refactor(agents): move claude's descriptor into its package (phase 2c)
Claude was the last agent whose command descriptor still lived in
`apps/cli/src/agents/` — 289 lines against codex's 18 and opencode's 11. Not
because claude is special, but because it is the only agent with `hooks`, and
those hooks called UP into the app: `runPrepare`, `maybeRunSetupWizard`,
`evaluateBaseFreshness`, `providerForBox`, the two clipboard paste helpers. An
agent package sits below `apps/cli`, so importing them would have closed an
`apps/cli -> package -> apps/cli` cycle.

They are passed in now, via `AgentHostServices` on the create context and a
`clipboard` argument to `attachExtras`. Nothing in the seam is claude-specific:
any agent that wants a setup wizard or image paste gets the same one.

Two things fell out of doing it properly rather than relocating the calls:

Three of the six collapsed into one. Claude never READ `baseStatus` — it graded
the base, decided whether a hub route made that grading irrelevant, ran the
wizard and re-baked on `rebuildBase`, all to forward a value it never inspected.
That is app policy end to end, so it lives behind a single `host.setupWizard()`.
`buildPromptArgs` needed no seam at all: it is a one-line wrapper over
`resolveAgentLauncher`, which agent packages already depend on.

The wizard's `command: 'create' | 'claude'` discriminant was never about claude
either — it separates the outer `agentbox create` pass from an agent's own pass.
Spelled `'agent'` now, so an agent reaching `host.setupWizard` gets the real
behavior instead of the `!== 'claude'` early return.

The guard is the part worth checking: it excluded `apps/cli/src/agents/`
wholesale, which meant it had never once looked at the descriptor this commit
moves. Narrowed to the four shim files, so the pipeline and both dispatch tables
are now measured like any other shared code — and it fails when an agent-named
export is planted in `agents/command/options.ts`.

Verified live on docker: create with a working claude session, `--plan` failing
fast on a bad path, and `--plan` on a real file landing the plan in the box and
launching `claude 'Resume the plan at …' --permission-mode plan`. Also checked
through a packed install, since the literal-specifier rule only fails there.
